move setWorkflowName
